moved OAuth strategies into subfolder, cleaned up some notes
