+++ src/OAuthStrategies/GoogleOAuth.ts
@@ -1,20 +1,21 @@
-// import { OAuthStrategyParams } from './bedrock.ts'
-import { Context, helpers } from "./deps.ts";
-import { GithubOAuthParams, Provider } from "./types.ts";
+import { Context, helpers } from "./../deps.ts";
+import { GoogleOAuthParams } from "./../types.ts";
 
-export class GithubOAuth {
-  provider = 'Github';
+export class GoogleOAuth {
+  provider = "Google";
   client_id: string;
   client_secret: string;
+  scope: string;
   redirect_uri: string;
+  response_type = "code";
+  access_type?: "online" | "offline";
   state?: string;
-  login?: string;
-  scope?: string;
-  allow_signup?: string;
+  prompt?: "none" | "consent" | "select_account";
 
-  constructor(stratParams: GithubOAuthParams) {
+  constructor(stratParams: GoogleOAuthParams) {
     this.client_id = stratParams.client_id;
     this.client_secret = stratParams.client_secret;
+    this.scope = stratParams.scope;
     this.redirect_uri = stratParams.redirect_uri;
     Object.assign(this, stratParams)!;
   }
@@ -23,10 +24,7 @@ export class GithubOAuth {
    * Appends client info onto uri string and redirects to generated link.
    */
   sendRedirect = (ctx: Context): void => {
-    let uri = "http://github.com/login/oauth/authorize?";
-    if (this.scope !== undefined) {
-      uri += `scope=${this.scope}&`;
-    }
+    let uri = "http://accounts.google.com/o/oauth2/v2/auth?";
     if (this.state === undefined) {
       this.state = "";
       const alphanum: string =
@@ -35,21 +33,27 @@ export class GithubOAuth {
         this.state += alphanum[Math.floor(Math.random() * alphanum.length)];
       }
     }
-
+    // Adding defined props to uri and including only props that are strings (exludes functions)
     for (let prop in this) {
       if (this[prop] !== undefined && prop !== "provider" && prop !== "client_secret" && typeof this[prop] === 'string') {
         uri += `${prop}=${this[prop]}&`;
       }
     }
+    // Removing trailing "&" from uri variable
+    uri = uri.slice(0, uri.length - 1);    
+    // console.log(uri);
     ctx.response.redirect(uri);
     return;
-  };
+  }
+
   /**
+   * After user signs in with their credentials on OAuth providers page a response 
+   * with the code and current state are sent back to client
    * 
    */
-  getToken = async (ctx: Context, next: () => Promise<unknown>) => {
-    const params = helpers.getQuery(ctx, { mergeParams: true });    
-    const { code, state } = params;    
+  getToken = async ( ctx: Context, next: () => Promise<unknown>) => {
+    const params = helpers.getQuery(ctx, { mergeParams: true });
+    const { code, state } = params;
 
     if (state !== this.state) {
       console.log("State validation on incoming response failed");
@@ -57,13 +61,13 @@ export class GithubOAuth {
       ctx.response.status = 401;
       ctx.response.body = {
         success: false,
-        message: "Unable to log in via Github",
+        message: "Unable to log in via Google",
       };
       throw new Error();
     }
 
     try {
-      const token = await fetch("https://github.com/login/oauth/access_token", {
+      const token = await fetch("https://oauth2.googleapis.com/token", {
         method: "POST",
         headers: {
           "Accept": "application/json",
@@ -73,6 +77,8 @@ export class GithubOAuth {
           client_id: this.client_id,
           client_secret: this.client_secret,
           code,
+          grant_type: "authorization_code",
+          redirect_uri: this.redirect_uri,
         }),
       });
 
@@ -90,9 +96,12 @@ export class GithubOAuth {
       };
       return;
     }
-  };
+  }
 
-  verifyAuth = async (ctx: Context, next: () => Promise<unknown>) => {
+  /**
+   * 
+   */
+  verifyAuth = async ( ctx: Context, next: () => Promise<unknown>) => {
     if (
       await ctx.state.session.has("isLoggedIn") &&
       await ctx.state.session.get("isLoggedIn")
@@ -108,10 +117,10 @@ export class GithubOAuth {
     }
     ctx.response.redirect("/blocked.html");
     return;
-  };
+  }
 
   signOut = async (ctx: Context, next: () => Promise<unknown>) => {
     await ctx.state.session.deleteSession(ctx);
     next();
-  };
+  }
 }
